Refactor tests and introduce conformance tests
This establishes test/conformance/, which contains tests which
exhaustively test each feature for comformance with the POSIX
specification down to the last detail. These tests should be useful for
verifying other shells as well, and do not require a test harness or a
reference shell (we still use both, though).

I'll DRY this up a bit once we support functions.