shell/task: create jobs with the pipeline/async list
Instead of creating jobs when executing their first process, create them when
creating the pipeline or async list.

This allows to track which pipeline created the job and will make possible to
attach the AST to the job in the future.

As of now, this breaks builtins such as fg(1) when called without an argument.
When executing fg(1) itself, a job is created and fg(1) picks this job instead
of the last one.