Make the shell responsible for destroying terminated jobs
Builtins call mrsh_run_program but don't want to destroy the current job
in the middle of the execution (which may be marked as terminated
because no process is running).

In the future, we'll probably want to let shells inspect the job list
before destroying terminated jobs (and maybe customise the notification
format).

Closes: https://github.com/emersion/mrsh/issues/122